1、基本完成增加微信小程序扫码登录；2、清理代码，修订部分machine报文格式；

+++ controller/preauth_keys.go
@@ -23,7 +23,7 @@ const (
 type PreAuthKey struct {
 	ID        uint64 `gorm:"primary_key"`
 	Key       string
-	UserID    uint
+	UserID    int64
 	User      User
 	Reusable  bool
 	Ephemeral bool `gorm:"default:false"`
